New Tests

+++ BasicCalculator/src/main/java/calculatorpkg/Calculator.java
@@ -13,14 +13,14 @@ package calculatorpkg;
  */
 public class Calculator {
 	
-	private double value1;
-	private double value2;
+	private int value1;
+	private int value2;
 	
 	/**
 	 * @param value1 holds the first number in the calculation
 	 * @param value2 holds the second number in the calculation
 	 */
-	public Calculator(double value1, double value2) {
+	public Calculator(int value1, int value2) {
 		super();
 		this.value1 = value1;
 		this.value2 = value2;
@@ -28,25 +28,25 @@ public class Calculator {
 	/**
 	 * @return the value1
 	 */
-	public double getValue1() {
+	public int getValue1() {
 		return value1;
 	}
 	/**
 	 * @param value1 the value1 to set
 	 */
-	public void setValue1(double value1) {
+	public void setValue1(int value1) {
 		this.value1 = value1;
 	}
 	/**
 	 * @return the value2
 	 */
-	public double getValue2() {
+	public int getValue2() {
 		return value2;
 	}
 	/**
 	 * @param value2 the value2 to set
 	 */
-	public void setValue2(double value2) {
+	public void setValue2(int value2) {
 		this.value2 = value2;
 	}
